+++ src/json-api-serializer.js
@@ -10,11 +10,14 @@ DS.JsonApiSerializer = DS.RESTSerializer.extend({
   relationshipKey: 'self',
   relatedResourceKey: 'related',
 
+  keyForAttribute: function(key) {
+    return Ember.String.dasherize(key);
+  },
   keyForRelationship: function(key) {
-    return key;
+    return Ember.String.dasherize(key);
   },
   keyForSnapshot: function(snapshot) {
-    return snapshot.typeKey;
+    return Ember.String.dasherize(snapshot.typeKey);
   },
 
   /**
@@ -179,7 +182,7 @@ DS.JsonApiSerializer = DS.RESTSerializer.extend({
   serialize: function(snapshot, options) {
     var data = this._super(snapshot, options);
     if(!data.hasOwnProperty('type') && options && options.type) {
-      data.type = Ember.String.pluralize(options.type);
+      data.type = Ember.String.pluralize(this.keyForRelationship(options.type));
     }
     return data;
   },
@@ -195,12 +198,11 @@ DS.JsonApiSerializer = DS.RESTSerializer.extend({
   },
 
   serializeIntoHash: function(hash, type, snapshot, options) {
-    var pluralType = Ember.String.pluralize(type.typeKey);
     var data = this.serialize(snapshot, options);
     if(!data.hasOwnProperty('type')) {
-      data.type = pluralType;
+      data.type = Ember.String.pluralize(this.keyForRelationship(type.typeKey));
     }
-    hash[type.typeKey] = data;
+    hash[this.keyForAttribute(type.typeKey)] = data;
   },
 
   /**